refactor the code to optimize project structure
