+++ .env-sample
@@ -1,0 +1,10 @@
+#kept key values as sample for testing. change the values accordingly in production
+
+SECRET_KEY = 'django-insecure-7+om-9__!v%1ud!6-wwkf0hs0x7v1myz9jn#e9d8n@v#1^qk3p'
+DEBUG=False
+
+DB_NAME=demo_db
+DB_USER=demo
+DB_PASSWORD=demo@123
+DB_HOST=DB_NAME
+DB_PORT=5432
