--- .env_sample
@@ -1,8 +1,0 @@
-SECRET_KEY = 'XXXXXX'
-DEBUG = False
-
-DB_NAME=demo_db
-DB_USER=demo
-DB_PASSWORD=demo@123
-DB_HOST=localhost
-DB_PORT=5432
