+++ Dockerfile
@@ -18,6 +18,9 @@ RUN pip install --no-cache-dir -r requirements.txt
 # Copy our codebase into the container
 COPY . .
 
+# Copy .env-sample as .env
+COPY .env-sample .env
+
 # RUN ./manage.py makemigrations
 # RUN ./manage.py migrate
 # RUN ./manage.py create_admin
