+++ api/models.py
@@ -1,47 +1,47 @@
-from django.db import models
-from django.contrib.auth.models import User
+# from django.db import models
+# from django.contrib.auth.models import User
 
-class Product(models.Model):
-    name = models.CharField(max_length=255)
-    description = models.TextField()
-    price = models.DecimalField(max_digits=10, decimal_places=2)
-    quantity = models.PositiveIntegerField()
+# class Product(models.Model):
+#     name = models.CharField(max_length=255)
+#     description = models.TextField()
+#     price = models.DecimalField(max_digits=10, decimal_places=2)
+#     quantity = models.PositiveIntegerField()
 
-    def __str__(self):
-        return self.name
+#     def __str__(self):
+#         return self.name
 
-class Cart(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE, null=True, blank=True)
-    cart_id = models.CharField(max_length=50, unique=True, blank=True, null=True)
+# class Cart(models.Model):
+#     user = models.ForeignKey(User, on_delete=models.CASCADE, null=True, blank=True)
+#     cart_id = models.CharField(max_length=50, unique=True, blank=True, null=True)
     
-    def __str__(self):
-        return f"{self.cart_id}"
+#     def __str__(self):
+#         return f"{self.cart_id}"
 
-class CartItem(models.Model):
-    cart = models.ForeignKey(Cart, on_delete=models.CASCADE)
-    product = models.ForeignKey(Product, on_delete=models.CASCADE)
-    quantity = models.PositiveIntegerField()
+# class CartItem(models.Model):
+#     cart = models.ForeignKey(Cart, on_delete=models.CASCADE)
+#     product = models.ForeignKey(Product, on_delete=models.CASCADE)
+#     quantity = models.PositiveIntegerField()
     
-    class Meta: #ensure that a product can only be added once
-        unique_together = ('cart', 'product')
+#     class Meta: #ensure that a product can only be added once
+#         unique_together = ('cart', 'product')
 
-    def __str__(self):
-        return f"{self.id}"
+#     def __str__(self):
+#         return f"{self.id}"
 
-class Order(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE, blank=True, null=True)
-    delivery_date_time = models.DateTimeField(blank=True, null=True)
-    name = models.CharField(max_length=255, blank=True, null=True)
-    email = models.EmailField(blank=True, null=True)
-    address = models.TextField(blank=True, null=True)
+# class Order(models.Model):
+#     user = models.ForeignKey(User, on_delete=models.CASCADE, blank=True, null=True)
+#     delivery_date_time = models.DateTimeField(blank=True, null=True)
+#     name = models.CharField(max_length=255, blank=True, null=True)
+#     email = models.EmailField(blank=True, null=True)
+#     address = models.TextField(blank=True, null=True)
 
-    def __str__(self):
-        return f"Order {self.id} for User {self.user_id}"
+#     def __str__(self):
+#         return f"Order {self.id} for User {self.user_id}"
 
-class OrderItem(models.Model):
-    order = models.ForeignKey(Order, on_delete=models.CASCADE)
-    product = models.ForeignKey(Product, on_delete=models.CASCADE)
-    quantity = models.PositiveIntegerField()
+# class OrderItem(models.Model):
+#     order = models.ForeignKey(Order, on_delete=models.CASCADE)
+#     product = models.ForeignKey(Product, on_delete=models.CASCADE)
+#     quantity = models.PositiveIntegerField()
 
-    def __str__(self):
-        return f"{self.quantity} of {self.product.name} in Order {self.order_id}"
+#     def __str__(self):
+#         return f"{self.quantity} of {self.product.name} in Order {self.order_id}"
