+++ cart/apps.py
@@ -1,0 +1,9 @@
+from django.apps import AppConfig
+
+
+class CartConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'cart'
+    
+    def ready(self):
+        import cart.signals
