+++ cart/serializers.py
@@ -1,0 +1,12 @@
+from rest_framework import serializers
+from cart.models import Cart, CartItem
+
+class CartSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Cart
+        fields = '__all__'
+
+class CartItemSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = CartItem
+        fields = '__all__'
