+++ cart/tests.py
@@ -1,5 +1,3 @@
-from django.test import TestCase
-
 from rest_framework.test import APITestCase
 from django.contrib.auth.models import User
 from rest_framework import status
