+++ product/views.py
@@ -1,0 +1,17 @@
+from rest_framework.viewsets import ModelViewSet
+from product.models import Product
+from .serializers import ProductSerializer
+from rest_framework.permissions import AllowAny
+
+from rest_framework.pagination import PageNumberPagination
+
+class CustomPagination(PageNumberPagination):
+    page_size = 5 
+    page_size_query_param = 'page_size'
+    max_page_size = 100
+
+class ProductViewSet(ModelViewSet):
+    queryset = Product.objects.all()
+    serializer_class = ProductSerializer
+    permission_classes = [AllowAny]
+    pagination_class = CustomPagination
